feat(db): category seed
