+++ apps/server/prisma/seeds/category.ts
@@ -1,0 +1,12 @@
+import { PrismaClient } from '@prisma/client';
+
+export const createCategory = async (prisma: PrismaClient) => {
+  for (let i = 0; i < 10; i++) {
+    await prisma.category.create({
+      data: {
+        name: `Category ${i}`,
+        kitchenId: 1
+      }
+    });
+  }
+};
